Refine maker editor architecture and bridge evaluation
